requirements.txt
torch  # ~= 1.11
dgl~=0.8
networkx~=2.6.3
pandas~=1.4.0
scikit-learn~=1.0
numpy~=1.21
numba~=0.55
# scipy~=1.6.3
pytz~=2021.3
matplotlib~=3.4.3
plotly
PyYAML~=5.4.1
pygraphviz
cdt~=0.5.23
loguru~=0.5.3
torchviz~=0.0.2
pyprof~=0.0.7
graphviz~=0.18.2
typed-argument-parser
pytorch-lightning
diskcache~=5.2.1
torchinfo~=1.5.3
tsfresh~=0.19.0
tap~=0.2
einops~=0.3.2
dill~=0.3.4
tqdm~=4.62.3
seaborn~=0.11.2
requests~=2.26.0
python-dateutil~=2.8.2
setuptools~=58.2.0
# pycausal==1.2.1
regex~=2021.11.10
lime~=0.2.0.1
ipython~=7.29.0
-e git+https://github.com/sarthak-chakraborty/CausIL.git@4f8adb188303fcae59eb68bf4fb49fc462efcca2#egg=causalDiscovery&subdirectory=LIB
